Refactors in date utils & basic filtering added
