+++ client/src/components/DateFilter.tsx
@@ -1,5 +1,13 @@
-import { useState } from "react";
-import { Menu, Button, Icon, Header, Popup, Form } from "semantic-ui-react";
+import { ChangeEvent, useState } from "react";
+import {
+  Menu,
+  Button,
+  Icon,
+  Header,
+  Popup,
+  Form,
+  Divider,
+} from "semantic-ui-react";
 import { currentViewDateRangeVar } from "../cache";
 import { ViewDateRange } from "../types";
 
@@ -38,6 +46,25 @@ const DateFilter = () => {
     setDateFilter(newDate);
   };
 
+  const handleDateFilterChange = (
+    event: ChangeEvent<HTMLInputElement>,
+    isStart: boolean
+  ) => {
+    console.log(event.target.value);
+    if (isStart) {
+      const newStart = new Date(event.target.value);
+      const newDate = { start: newStart, end: dateFilter.end };
+      currentViewDateRangeVar(newDate);
+      setDateFilter(newDate);
+    } else {
+      const newEnd = new Date(event.target.value);
+      const newDate = { start: dateFilter.start, end: newEnd };
+      currentViewDateRangeVar(newDate);
+      setDateFilter(newDate);
+    }
+    console.log(dateFilter);
+  };
+
   const resetMonth = () => {
     currentViewDateRangeVar(initDate);
     setDateFilter(initDate);
@@ -62,16 +89,24 @@ const DateFilter = () => {
               </Header>
             </Button>
           }
+          position="top center"
         >
           <Form>
             <Form.Input
               label="Start"
               type="date"
-              value={dateFilter.start.toString()}
+              onChange={(event) => handleDateFilterChange(event, true)}
+              //value={getYearMonthDay(dateFilter.start)}
             />
-            <Form.Input label="End" type="date" value={dateFilter.end} />
+            <Form.Input
+              label="End"
+              type="date"
+              onChange={(event) => handleDateFilterChange(event, false)}
+              //value={getYearMonthDay(dateFilter.end)}
+            />
           </Form>
-          <Button onClick={() => resetMonth()}>View ongoing</Button>
+          <Divider />
+          <Button onClick={() => resetMonth()}>Reset</Button>
         </Popup>
         <Menu.Item fitted>
           <Button attached="right" size="mini" onClick={() => nextMonth()}>
