+++ client/src/components/EntryForm.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import * as Yup from "yup";
 import { FormikProps, Field } from "formik";
 import { useState, useEffect } from "react";
